+++ phison_realestate_mobile/lib/shared/widgets/error_message.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class ErrorMessage extends StatelessWidget {
+  const ErrorMessage({
+    Key? key,
+    required this.error,
+  }) : super(key: key);
+
+  final String error;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(error),
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).popUntil((route) => route.isFirst);
+            },
+            child: const Text('Go to Login'),
+          )
+        ],
+      ),
+    );
+  }
+}
