--- phison_realestate_mobile/lib/shared/widgets/error_text.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ErrorText extends StatelessWidget {
-  const ErrorText({
-    Key? key,
-    required this.error,
-  }) : super(key: key);
-
-  final String error;
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Text(error),
-    );
-  }
-}
